added readme file and updated the comment

+++ src/server/priority-request-generator-server/PriorityRequestGeneratorServer.h
@@ -54,7 +54,7 @@ public:
     bool checkSrmSendingFlag();
     int getMessageType(string jsonString);
     int getTimedOutVehicleID();
-    int getCurrentTimeInSeconds();
+    double getCurrentTimeInSeconds();
     double haversineDistance(double lat1, double lon1, double lat2, double lon2);
     string getSRMJsonString();
     string getPrgStatusJsonString();
